Update partial quest progress

+++ contracts/globals/actions.sol
@@ -87,7 +87,7 @@ struct Quest {
   uint16 rewardAmount; // amount of the reward (up to 65535)
   uint16 rewardItemTokenId1; // Reward another item
   uint16 rewardAmount1; // amount of the reward (up to 65535)
-  uint56 questId; // Unique id for this quest
+  uint40 questId; // Unique id for this quest
 }
 
 struct QuestWithCompletionInfo {
@@ -95,5 +95,5 @@ struct QuestWithCompletionInfo {
   uint24 actionNum;
   uint24 actionNum1;
   uint24 actionChoiceNum;
-  uint56 realQuestId; // If daily this is incremented to differentiate the same quest
+  uint40 realQuestId; // If daily this is incremented to differentiate the same quest
 }
